moving DataTable to its own class
